+++ DEWrapperWorkflow/src/main/java/io/seqware/pancancer/JobUtilities.java
@@ -64,14 +64,15 @@ public class JobUtilities {
   public Job localUploadJob(Job uploadJob, String workflowDataDir, String pemFile, String metadataURLs,
           List<String> vcfs, List<String> vcfmd5s, List<String> tbis, List<String> tbimd5s,
           List<String> tars, List<String> tarmd5s, String uploadServer, String seqwareVersion,
-          String vmInstanceType, String vmLocationCode, String overrideTxt, String uploadLocalPath, String temp, int timeout, int retries) {
+          String vmInstanceType, String vmLocationCode, String overrideTxt, String uploadLocalPath, String temp,
+          int timeout, int retries, String qcJson, String timingJson) {
     
     StringBuffer sb = new StringBuffer(overrideTxt);
     sb.append(" --upload-archive "+temp+" --skip-upload --skip-validate ");
     
     uploadJob = vcfUpload(uploadJob, workflowDataDir, pemFile, metadataURLs,
           vcfs, vcfmd5s, tbis, tbimd5s, tars, tarmd5s, uploadServer, seqwareVersion,
-          vmInstanceType, vmLocationCode, sb.toString(), timeout, retries);
+          vmInstanceType, vmLocationCode, sb.toString(), timeout, retries, qcJson, timingJson);
     
     uploadJob.getCommand().addArgument(" && rsync -rauv "+temp+"/*.tar.gz "+uploadLocalPath+"/");
     
@@ -82,11 +83,12 @@ public class JobUtilities {
   public Job gnosUploadJob(Job uploadJob, String workflowDataDir, String pemFile, String metadataURLs,
           List<String> vcfs, List<String> vcfmd5s, List<String> tbis, List<String> tbimd5s,
           List<String> tars, List<String> tarmd5s, String uploadServer, String seqwareVersion,
-          String vmInstanceType, String vmLocationCode, String overrideTxt, int timeout, int retries) {
+          String vmInstanceType, String vmLocationCode, String overrideTxt, int timeout, int retries,
+          String qcJson, String timingJson) {
     
     return(vcfUpload(uploadJob, workflowDataDir, pemFile, metadataURLs,
           vcfs, vcfmd5s, tbis, tbimd5s, tars, tarmd5s, uploadServer, seqwareVersion,
-          vmInstanceType, vmLocationCode, overrideTxt, timeout, retries));
+          vmInstanceType, vmLocationCode, overrideTxt, timeout, retries, qcJson, timingJson));
     
   }
   
@@ -97,14 +99,15 @@ public class JobUtilities {
           List<String> vcfs, List<String> vcfmd5s, List<String> tbis, List<String> tbimd5s,
           List<String> tars, List<String> tarmd5s, String uploadServer, String seqwareVersion,
           String vmInstanceType, String vmLocationCode, String overrideTxt, String temp,
-          String S3UploadArchiveKey, String S3UploadArchiveSecretKey, String uploadS3Bucket, int timeout, int retries) {
+          String S3UploadArchiveKey, String S3UploadArchiveSecretKey, String uploadS3Bucket,
+          int timeout, int retries, String qcJson, String timingJson) {
 
       StringBuffer sb = new StringBuffer(overrideTxt);
       sb.append(" --upload-archive "+temp+" --skip-upload --skip-validate ");
     
       uploadJob = vcfUpload(uploadJob, workflowDataDir, pemFile, metadataURLs,
           vcfs, vcfmd5s, tbis, tbimd5s, tars, tarmd5s, uploadServer, seqwareVersion,
-          vmInstanceType, vmLocationCode, sb.toString(), timeout, retries);
+          vmInstanceType, vmLocationCode, sb.toString(), timeout, retries, qcJson, timingJson);
     
       uploadJob.getCommand()
         .addArgument(" && mkdir -p ~/.aws/; ")
@@ -122,7 +125,8 @@ public class JobUtilities {
   public Job vcfUpload(Job uploadJob, String workflowDataDir, String pemFile, String metadataURLs,
           List<String> vcfs, List<String> vcfmd5s, List<String> tbis, List<String> tbimd5s,
           List<String> tars, List<String> tarmd5s, String uploadServer, String seqwareVersion,
-          String vmInstanceType, String vmLocationCode, String overrideTxt, int timeout, int retries) {
+          String vmInstanceType, String vmLocationCode, String overrideTxt, int timeout, int retries,
+          String qcJson, String timingJson) {
     
     uploadJob.getCommand().addArgument(
                 "docker run "
@@ -135,7 +139,7 @@ public class JobUtilities {
                         + ":/root/gnos_icgc_keyfile.pem "
                         + "seqware/pancancer_upload_download "
                         // the command invoked on the container follows
-                        + "/bin/bash -c 'cd /workflow_data && echo '{}' > /tmp/empty.json && mkdir -p uploads && "
+                        + "/bin/bash -c 'cd /workflow_data && mkdir -p uploads && "
                         + "perl -I /opt/gt-download-upload-wrapper/gt-download-upload-wrapper-2.0.7/lib "
                         + "/opt/vcf-uploader/vcf-uploader-2.0.2/gnos_upload_vcf.pl "
                         // parameters to gnos_upload
@@ -145,14 +149,15 @@ public class JobUtilities {
                         + Joiner.on(',').join(tbis) + " --vcf-idx-md5sum-files " + Joiner.on(',').join(tbimd5s) + " --tarballs "
                         + Joiner.on(',').join(tars) + " --tarball-md5sum-files " + Joiner.on(',').join(tarmd5s) + " --outdir uploads" 
                         + " --key /root/gnos_icgc_keyfile.pem --upload-url " + uploadServer
-                        + " --qc-metrics-json /tmp/empty.json" + " --timing-metrics-json /tmp/empty.json"
+                        + " --qc-metrics-json " + qcJson
+                        + " --timing-metrics-json " + timingJson
                         + " --workflow-src-url https://bitbucket.org/weischen/pcawg-delly-workflow" + "--workflow-url https://registry.hub.docker.com/u/pancancer/pcawg-delly-workflow" + " --workflow-name EmblPancancerStr "
                         + " --timeout-min "+timeout+" --retries "+retries+" "
                         + " --workflow-version 1.0.0" + " --seqware-version " + seqwareVersion + " --vm-instance-type "
                         + vmInstanceType + " --vm-instance-cores `nproc` --vm-instance-mem-gb "
-                        + "`free | grep 'Mem:' | awk '{print $2 / 1000000 }'` " 
+                        + "`free | grep \"Mem:\" | awk \"{print \\\\$2 / 1000000 }\"` "
                         + " --vm-location-code " + vmLocationCode + overrideTxt
-                        + " --timeout-min "+timeout+" --retries "+retries+" "
+                        + " --timeout-min "+timeout+" --retries "+retries+"' "
                         );
     
     return(uploadJob);